feat: add comprehensive tests for committee resolution and proof construction in WASM and Rust implementations
