+++ poi-rs/tests/proof_construction.rs
@@ -5,159 +5,14 @@ mod utils;
 
 use std::sync::{Arc, Mutex};
 
-use async_trait::async_trait;
-use iota_grpc_client::Client as GrpcClient;
-use iota_sdk_types::{ObjectId, TransactionDigest, Version};
-use iota_types::{
-    committee::{Committee, EpochId},
-    digests::ChainIdentifier,
-    event::EventID,
-    messages_checkpoint::CertifiedCheckpointSummary,
-    object::Object,
+use iota_sdk_types::TransactionDigest;
+use iota_types::{event::EventID, object::Object};
+use poi_rs::{PoiClient, ProofBuilderError, ProofTarget, ProofVerifier, SourceError};
+use utils::{
+    genesis_chain_identifier, grpc_client, object_transfer_tx,
+    sources::{MissingSource, RecordingSource, RejectingSource},
+    staking_tx, start_test_cluster, transfer_tx,
 };
-use poi_rs::{
-    PoiClient, ProofBuilderError, ProofTarget, ProofVerifier, Source, SourceCheckpoint, SourceError, SourceTransaction,
-};
-use utils::{genesis_chain_identifier, grpc_client, object_transfer_tx, staking_tx, start_test_cluster, transfer_tx};
-
-#[derive(Clone)]
-struct RejectingSource;
-
-#[async_trait]
-impl Source for RejectingSource {
-    async fn chain_identifier(&self) -> Result<ChainIdentifier, SourceError> {
-        unreachable!("rejected transactions do not resolve a chain identifier")
-    }
-
-    async fn transaction(
-        &self,
-        _transaction_digest: TransactionDigest,
-    ) -> Result<Option<SourceTransaction>, SourceError> {
-        Err(SourceError::request(std::io::Error::other("transaction rejected")))
-    }
-
-    async fn object(&self, _object_id: ObjectId, _version: Option<Version>) -> Result<Option<Object>, SourceError> {
-        Ok(None)
-    }
-
-    async fn checkpoint(&self, _sequence_number: u64) -> Result<SourceCheckpoint, SourceError> {
-        unreachable!("rejected transactions do not resolve a checkpoint")
-    }
-
-    async fn committee(&self, _epoch: EpochId) -> Result<Committee, SourceError> {
-        unreachable!("proof-only test source does not resolve committees")
-    }
-
-    async fn current_epoch(&self) -> Result<Option<EpochId>, SourceError> {
-        unreachable!("proof-only test source does not resolve the current epoch")
-    }
-
-    async fn epoch_close_summary(&self, _epoch: EpochId) -> Result<Option<CertifiedCheckpointSummary>, SourceError> {
-        unreachable!("proof-only test source does not resolve epoch-close summaries")
-    }
-}
-
-#[derive(Clone)]
-struct RecordingSource {
-    source: GrpcClient,
-    transactions: Arc<Mutex<Vec<TransactionDigest>>>,
-    object_override: Option<Object>,
-}
-
-impl RecordingSource {
-    fn new(source: GrpcClient, transactions: Arc<Mutex<Vec<TransactionDigest>>>) -> Self {
-        Self {
-            source,
-            transactions,
-            object_override: None,
-        }
-    }
-
-    fn with_object_override(mut self, object: Object) -> Self {
-        self.object_override = Some(object);
-        self
-    }
-}
-
-#[async_trait]
-impl Source for RecordingSource {
-    async fn chain_identifier(&self) -> Result<ChainIdentifier, SourceError> {
-        self.source.chain_identifier().await
-    }
-
-    async fn transaction(
-        &self,
-        transaction_digest: TransactionDigest,
-    ) -> Result<Option<SourceTransaction>, SourceError> {
-        self.transactions
-            .lock()
-            .expect("recorded transactions lock must not be poisoned")
-            .push(transaction_digest);
-
-        self.source.transaction(transaction_digest).await
-    }
-
-    async fn object(&self, object_id: ObjectId, version: Option<Version>) -> Result<Option<Object>, SourceError> {
-        if let Some(object) = &self.object_override {
-            return Ok(Some(object.clone()));
-        }
-
-        self.source.object(object_id, version).await
-    }
-
-    async fn checkpoint(&self, sequence_number: u64) -> Result<SourceCheckpoint, SourceError> {
-        self.source.checkpoint(sequence_number).await
-    }
-
-    async fn committee(&self, epoch: EpochId) -> Result<Committee, SourceError> {
-        self.source.committee(epoch).await
-    }
-
-    async fn current_epoch(&self) -> Result<Option<EpochId>, SourceError> {
-        self.source.current_epoch().await
-    }
-
-    async fn epoch_close_summary(&self, epoch: EpochId) -> Result<Option<CertifiedCheckpointSummary>, SourceError> {
-        self.source.epoch_close_summary(epoch).await
-    }
-}
-
-#[derive(Clone)]
-struct MissingSource;
-
-#[async_trait]
-impl Source for MissingSource {
-    async fn chain_identifier(&self) -> Result<ChainIdentifier, SourceError> {
-        unreachable!("missing targets do not resolve a chain identifier")
-    }
-
-    async fn transaction(
-        &self,
-        _transaction_digest: TransactionDigest,
-    ) -> Result<Option<SourceTransaction>, SourceError> {
-        Ok(None)
-    }
-
-    async fn object(&self, _object_id: ObjectId, _version: Option<Version>) -> Result<Option<Object>, SourceError> {
-        Ok(None)
-    }
-
-    async fn checkpoint(&self, _sequence_number: u64) -> Result<SourceCheckpoint, SourceError> {
-        unreachable!("missing targets do not resolve a checkpoint")
-    }
-
-    async fn committee(&self, _epoch: EpochId) -> Result<Committee, SourceError> {
-        unreachable!("proof-only test source does not resolve committees")
-    }
-
-    async fn current_epoch(&self) -> Result<Option<EpochId>, SourceError> {
-        unreachable!("proof-only test source does not resolve the current epoch")
-    }
-
-    async fn epoch_close_summary(&self, _epoch: EpochId) -> Result<Option<CertifiedCheckpointSummary>, SourceError> {
-        unreachable!("proof-only test source does not resolve epoch-close summaries")
-    }
-}
 
 #[tokio::test]
 async fn client_uses_a_custom_source_for_proof_building() {
